# openapi3.yaml
openapi: 3.0.1
info:
  title: Geospatial Products Catalog
  description: basic template for map colonies service for managing geospatial products
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: /
    description: Current Environment

security:
  - {}

paths:
  /products:
    get:
      operationId: getProducts
      tags:
        - Product
      summary: List products with geospatial filters
      parameters:
        - $ref: '#/components/parameters/Name'
        - $ref: '#/components/parameters/ProductType'
        - $ref: '#/components/parameters/ConsumptionProtocol'
        - $ref: '#/components/parameters/MinZoom'
        - $ref: '#/components/parameters/MaxZoom'
        - $ref: '#/components/parameters/ResolutionBest'
        - $ref: '#/components/parameters/MinZoomGreater'
        - $ref: '#/components/parameters/MaxZoomGreater'
        - $ref: '#/components/parameters/BoundingPolygonIntersects'
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

    post:
      operationId: createProduct
      tags:
        - Product
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        201:
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

  /products/{id}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      operationId: getProductById
      tags:
        - Product
      summary: Get product by id
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateProduct
      tags:
        - Product
      summary: Partially update product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteProduct
      tags:
        - Product
      summary: Delete product
      responses:
        204:
          description: Deleted
        404:
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: integer

    Name:
      name: name
      in: query
      schema:
        type: string
    ProductType:
      name: type
      in: query
      schema:
        $ref: '#/components/schemas/ProductType'
    ConsumptionProtocol:
      name: consumptionProtocol
      in: query
      schema:
        $ref: '#/components/schemas/ConsumptionProtocol'
    MinZoom:
      name: minZoom
      in: query
      schema:
        $ref: '#/components/schemas/ZoomLevel'
    MaxZoom:
      name: maxZoom
      in: query
      schema:
        $ref: '#/components/schemas/ZoomLevel'
    ResolutionBest:
      name: resolutionBest
      in: query
      schema:
        $ref: '#/components/schemas/Resolution'
    MinZoomGreater:
      name: minZoomGreater
      in: query
      schema:
        $ref: '#/components/schemas/ZoomLevel'
    MaxZoomGreater:
      name: maxZoomGreater
      in: query
      schema:
        $ref: '#/components/schemas/ZoomLevel'
    BoundingPolygonIntersects:
      name: boundingPolygonIntersects
      in: query
      description: Geometry in WKT (Well-Known Text) format
      schema:
        $ref: '#/components/schemas/GeometryString'

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'

  schemas:
    error:
      type: object
      required: [message]
      properties:
        message:
          type: string

    ProductType:
      type: string
      enum: [raster, rasterized vector, 3d tiles, QMesh]

    ConsumptionProtocol:
      type: string
      enum: [WMS, WMTS, XYZ, 3D Tiles]

    ZoomLevel:
      type: integer
      minimum: 0
      maximum: 24

    Resolution:
      type: number
      format: double

    GeometryString:
      type: string

    ProductInput:
      type: object
      required: [name, boundingPolygon, type, consumptionProtocol]
      properties:
        name:
          type: string
        description:
          type: string
        boundingPolygon:
          $ref: '#/components/schemas/GeometryString'
        type:
          $ref: '#/components/schemas/ProductType'
        consumptionProtocol:
          $ref: '#/components/schemas/ConsumptionProtocol'
        resolutionBest:
          $ref: '#/components/schemas/Resolution'
        minZoom:
          $ref: '#/components/schemas/ZoomLevel'
        maxZoom:
          $ref: '#/components/schemas/ZoomLevel'

    ProductUpdate:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        description:
          type: string
        boundingPolygon:
          $ref: '#/components/schemas/GeometryString'
        type:
          $ref: '#/components/schemas/ProductType'
        consumptionProtocol:
          $ref: '#/components/schemas/ConsumptionProtocol'
        resolutionBest:
          $ref: '#/components/schemas/Resolution'
        minZoom:
          $ref: '#/components/schemas/ZoomLevel'
        maxZoom:
          $ref: '#/components/schemas/ZoomLevel'

    Product:
      type: object
      required: [id, name, type, consumptionProtocol, boundingPolygon]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
        description:
          type: string
        boundingPolygon:
          type: object
          description: GeoJSON representation of the polygon (as returned by DB)
        type:
          $ref: '#/components/schemas/ProductType'
        consumptionProtocol:
          $ref: '#/components/schemas/ConsumptionProtocol'
        resolutionBest:
          $ref: '#/components/schemas/Resolution'
        minZoom:
          $ref: '#/components/schemas/ZoomLevel'
        maxZoom:
          $ref: '#/components/schemas/ZoomLevel'